requirements.txt
fastapi
uvicorn
pinecone-client
python-dotenv
google-generativeai
pydantic
pydantic-settings
